Predicción Proteica
The primary structure of a protein refers to the linear sequence of amino acids that make up the protein, and determining this sequence is critical to understanding its function and biological behavior. The process of predicting the primary structure of proteins is a critical challenge in bioinformatics and computational biology.